separate globals into its own class
